write policies sparsely
Policy targets are ~1.5% nonzero, so the dense float32[1858] was almost
entirely zeros. New records store (int16 indices, float32 values); readers
already accept either form, so old shards keep working and no migration pass
is needed.

Sparsified at the append sites rather than at write time, so the live buffer
holds the compact form too -- at LIVE_BUFFER_SIZE=163840 that is ~3.6 GB of
policy+mask down to ~64 MB.

Also drop the int32 cast in records_to_arrays and the one in the tfrec read.
x is int16 everywhere and torch recasts on the way to the GPU, so the int32
intermediate only doubled the array.

Verified end to end on 102400 real records: write through the production
path, read back, every policy bit-identical. 28.0 MB -> 18.6 MB.

+++ src/chessbot/retrain_worker.py
@@ -81,7 +81,10 @@ def predict_fp16(model, X_list, batch_size):
 
 
 def records_to_arrays(records):
-    X = np.stack([r[0] for r in records]).astype(np.int32)
+    # x is int16 on every path and torch recasts on the way to the GPU
+    # (.long() for token encodings, .float() for lc0 planes), so an int32
+    # intermediate here just doubles the array for nothing
+    X = np.stack([r[0] for r in records])
     # primary + replay + historic meet here, so this is where sparse (new)
     # and dense (old, and historic tfrec) records mix
     P = stack_policies(records)
